SpeciesDistributionModels

Documentation for SpeciesDistributionModels.

SpeciesDistributionModels.remove_collinear

SpeciesDistributionModels.remove_collinear — Method

remove_collinear(data; method, silent = false)

Removes strongly correlated variables in data, until correlation is below a threshold specified in method.

method can currently be either Gvif, Vif or Pearson, which use GVIF, VIF, or Pearson's r, respectively. GVIF and VIF are similar method, but GVIF includes categorical variables whereas VIF ignores them.

To run without showing information about collinearity scores, set silent = true.

Example

julia> import SpeciesDistributionModels as SDM
julia> mydata = (a = 1:100, b = sqrt.(1:100), c = rand(100))
julia> SDM.remove_collinear(mydata; method = SDM.Vif(10))
[ Info: a has highest GVIF of 28.367942095054225
[ Info: Removing a, 2 variables remaining
[ Info: b has highest GVIF of 1.0077618445543057
[ Info: All variables are below threshold, returning remaining variables
(:b, :c)