SpeciesDistributionModels

Documentation for SpeciesDistributionModels.

SpeciesDistributionModels.ShapleyValues

SpeciesDistributionModels.evaluate

SpeciesDistributionModels.explain

SpeciesDistributionModels.machine_evaluations

SpeciesDistributionModels.predict

SpeciesDistributionModels.remove_collinear

SpeciesDistributionModels.sdm

SpeciesDistributionModels.ShapleyValues — Type

ShapleyValues(algorithm::Shapley.Algorithm)
ShapleyValues(N::Integer; threaded = true, rng = Random.GLOBAL_RNG)

Use to specify use Shapley values as method in explain. If an integer N, and optionally threaded and rng is supplied, MonteCarlo sampling is used, where N is the number of iterations (samples). More samples will result in more accurate results, but will take more time to compute.

SpeciesDistributionModels.evaluate — Method

evaluate(x; measures, train = true, test = true, [validation])

Evaluate x, which could be a SDMmachine, SDMgroup, or SDMensemble, by applying the measures provided to the data used to built an ensemble, and return an evaluation object.

Keywords

measures is a NamedTuple of measures. The keys are used to identify the measures.

This defaults to using auc, logloss, and kappa. For threshold-dependent measures, the highest score as well as the threshold at which the highest scores is reached are reported. A list of measurse is available here: https://juliaai.github.io/StatisticalMeasures.jl/dev/autogeneratedlistof_measures/#aliases. However, note that not all measures are useful.

train: a Bool that indicates whether to evaluate on the training data. Defaults to true

test: a Bool that indicates whether to evaluate on the test data.Defaults to true

validation: optionally, provide a separate validation dataset.

Data should be provided as a Tuple with presences as the first field and absences as the second.

SpeciesDistributionModels.explain — Method

explain(ensemble::SDMensemble; method, [data], [predictors])

Generate response curves for ensemble.

Keywords

method is the algorithm to use. See ShapleyValues

data is the data to use to generate response curves, and defaults to the data used to train the ensemble

predictors: which predictors to generate response curves for. Defaults to all variables in data.

SpeciesDistributionModels.machine_evaluations — Function

machine_evaluations(eval)

Get the scores for each machine in an evaluation, which can be either an 
`SDMgroupEvaluation` or an `SDMensembleEvaluation`.

The return type is a nested structure of `NamedTuple`s. 
The `NamedTuple` returned has keys corresponding to the evaluation datasets (a subset of :train, :test, and :validation), 
which each have keys corresponding to the measures specified in [`evaluate`](@ref).

## Example
```julia
evaluation = SDM.evaluate(ensemble; measures = (; accuracy, auc))
machine_aucs = SDM.machine_evaluations(evaluation).train.auc
```

SpeciesDistributionModels.predict — Method

predict(SDMobject, newdata; clamp = false, threaded = true, [reducer], [by_group])

Use an SDMmachine, SDMgroup, or SDMensemble to predict habitat suitability for some data, optionally summarized for the entire ensemble, or for each SDMgroup.

newdata can be either a RasterStack, or a Tables.jl.compatible object. It must have all predictor variables used to train the models in its columns (or layers in case of a RasterStack).

Keywords

clamp: if true, the predictions are clamped to the interval seen during training of SDMobject. Defaults to false

threaded: if true, run multithreaded. Defaults to true.

reducer: Optionally provide a Function to summarize the output. The function should take an vector of values and return a single value. Typical examples are Statistics.mean or Statistics.median.

by_group is set to true, the data is reduced for each SDMgroup, if it is set to false (the default), it reduced across the entire ensemble.

Returns

If newdata is a RasterStack, the predict returns a Raster; otherwise, it returns a NamedTuple of Vectors, with habitat suitability represented by a floating-point number between 0 and 1.

SpeciesDistributionModels.remove_collinear — Method

remove_collinear(data; method, silent = false)

Removes strongly correlated variables in data, until correlation is below a threshold specified in method.

method can currently be either Gvif, Vif or Pearson, which use GVIF, VIF, or Pearson's r, respectively. GVIF and VIF are similar method, but GVIF includes categorical variables whereas VIF ignores them.

To run without showing information about collinearity scores, set silent = true.

Example

julia> import SpeciesDistributionModels as SDM
julia> mydata = (a = 1:100, b = sqrt.(1:100), c = rand(100))
julia> SDM.remove_collinear(mydata; method = SDM.Vif(10))
[ Info: a has highest GVIF of 28.367942095054225
[ Info: Removing a, 2 variables remaining
[ Info: b has highest GVIF of 1.0077618445543057
[ Info: All variables are below threshold, returning remaining variables
(:b, :c)

SpeciesDistributionModels.sdm — Method

sdm(presences, absences; models, [resampler], [predictors], [verbosity])

Construct an ensemble with input data specified in presences and absences.

The first input argument is species presences and the second (pseudo-)absences. Both presence and absence data must be Tables-compatible (e.g., a DataFrame, a Vector of NamedTuple, but not an Array)

Keywords

models: a Vector of the models to be used in the ensemble. All models must be MLJ-supported Classifiers. For a full list of supported models, see https://alan-turing-institute.github.io/MLJ.jl/stable/model_browser/#Classification resampler: The resampling strategy to be used of type MLJBase.ResamplingStrategy. Defaults to 5-fold cross validation. predictors: a Vector of Symbols with the names of the predictor values to be used. By default, all pdf verbosity: an Int value that regulates how much information is printed.

Example